feat: split out and export sending feedback data
